moment component broken out into wrapper, card, and details
